complete delete option for each exercise in workout

+++ SpringFit/src/main/java/com/skilldistillery/springfit/controllers/WorkoutController.java
@@ -286,9 +286,7 @@ public class WorkoutController {
 			@RequestParam("workoutId") Integer workoutId) {
 		ModelAndView mv = new ModelAndView();
 		try {
-//			for (WorkoutExercise workoutExercise : workout.getWorkoutExercises()) {
 			workoutExerciseDao.updateWorkoutExercise(workoutExerciseId, workoutExercise);
-//			}
 			mv.setViewName("redirect:updateWorkoutExerciseForm.do?workoutId=" + workoutId);
 		} catch (Exception e) {
 			mv.addObject("errorMsg", "Error occurred while updating the workout exercises.");
@@ -297,55 +295,30 @@ public class WorkoutController {
 		}
 		return mv;
 	}
-	
-	@RequestMapping(path = "deleteWorkoutExercise.do", method = RequestMethod.POST)
-	public ModelAndView deleteWorkoutExercise(@ModelAttribute("workout") WorkoutExercise workoutExercise, @RequestParam("workoutExerciseId") Integer workoutExerciseId, 
-			@RequestParam("workoutId") Integer workoutId) {
-		
-		System.out.println("WorkoutExerciseId: " + workoutExerciseId);
-	    System.out.println("WorkoutId: " + workoutId);
-		ModelAndView mv = new ModelAndView();
-		
-		try {
-			boolean deleted = workoutExerciseDao.deleteWorkoutExercise(workoutExerciseId);
-			if (deleted) {
-				mv.setViewName("redirect:updateWorkoutExerciseForm.do?workoutId=" + workoutId);
-			} else {
-				mv.addObject("errorMsg", "Failed to delete WorkoutExercise with ID: " + workoutExerciseId);
-				mv.setViewName("error");
-			}
-		} catch (Exception e) {
-			mv.addObject("errorMsg", "Cannot delete WorkoutExercise with ID: " + workoutExerciseId + " due to foreign key constraints.");
-			mv.setViewName("error");
-			e.printStackTrace();
-		}
-		
-		return mv;
-	}
 
-//	@RequestMapping(path = "deleteWorkoutExercise.do", method = RequestMethod.POST)
-//	public ModelAndView deleteWorkoutExercise(@ModelAttribute("workout") WorkoutExercise workoutExercise,
-//			@RequestParam("workoutExerciseId") Integer workoutExerciseId, 
-//			@RequestParam("workoutId") Integer workoutId) {
-//
-//		ModelAndView mv = new ModelAndView();
-////		workoutExercise = workoutExerciseDao.getWorkoutExerciseById(workoutExerciseId);
-//
-//		if (workoutExercise == null) {
-//			mv.addObject("errorMsg", "WorkoutExercise not found with ID: " + workoutExerciseId);
-//			mv.setViewName("error");
-//		} else {
-//			boolean deleted = workoutExerciseDao.deleteWorkoutExercise(workoutExerciseId);
-//			if (!deleted) {
-//				mv.addObject("errorMsg", "Failed to delete");
-//				mv.setViewName("error");
-//			}else {
-//				mv.setViewName("redirect:updateWorkoutExerciseForm.do?workoutId=" + workoutId);
-//			}
-//		}
-//		return mv;
-//
-//	}
+
+	@RequestMapping(path = "deleteWorkoutExercise.do", method = RequestMethod.POST)
+	public ModelAndView deleteWorkoutExercise(@ModelAttribute("workout") WorkoutExercise workoutExercise,
+			@RequestParam("workoutExerciseId") Integer workoutExerciseId, 
+			@RequestParam("workoutId") Integer workoutId) {
+
+		ModelAndView mv = new ModelAndView();
+
+		if (workoutExercise == null) {
+			mv.addObject("errorMsg", "WorkoutExercise not found with ID: " + workoutExerciseId);
+			mv.setViewName("error");
+		} else {
+			boolean deleted = workoutExerciseDao.deleteWorkoutExercise(workoutExerciseId);
+			if (!deleted) {
+				mv.addObject("errorMsg", "Failed to delete");
+				mv.setViewName("error");
+			}else {
+				mv.setViewName("redirect:updateWorkoutExerciseForm.do?workoutId=" + workoutId);
+			}
+		}
+		return mv;
+
+	}
